Automate setting OMG_VERSION
This script will set the OMG_VERSION to a string containing the build
env name, OMG Release number, gihthub release tag and git branch.  And
to invoke it needs to be added as an extra script in the env section of
platform.ini

extra_scripts = pre:scripts/omg_firmware_version.py

I.e.

esp32dev_cccccc-v0.9.6-12-gdea42b3[development-switching]

Tested with MacOS, will need validation with a windows environment.  If
it fails validation on windows am thinking to just comment it out in
platformio.ini and add a comment indicating the feature and optionally
enable.