Process BLE advertisements on core 1 / reinit BLE to avoid WDT timeout
This creates a FreeRTOS queue and task on core 1 to process incoming BLE scan results.
The purpose is to free the core 0 cpu from the extra processing and potentially improve performance.

This also adds back the previously removed re-initializing of the BLE stack on each scan as a workaround for a BLE controller bug
that causes a watchdog timer timeout error after some time.
This should be considered for removal in the future if/when the bug is corrected.

Adds tracing for the 2 task stack high watermarks.