Modify home assistant configuration changes to reflect MQTT syntax that is required since HomeAssistant 2022.12.0
The MQTT YAML configuration change was optional in 2022.6, but the new syntax (updated above) is now mandatory, and now matches the already updated documentation provided at the top of the page.  I also added the force_update flag to help with the expire_after to give better graphing of temperature/humidity readings.