add '-DZsensorTEMT6000="TEMT6000"' to esp32dev-all-test
We should not add three other environments. The users should store this kind of custom environment in a prod_env.ini
By adding the module into the nodemcu-all-test and the esp32dev-all-test we verify that the module builds correctly and can be used by changing the configuration of the gateway through Platformio.