Fix connection validation on unknown indexes
Connection with Index 1 and 2 were supposed valid when not present in the configuration, generating attempts to connect with those indexes.
And improve erase config by adding an erase when the config file is not present (ESP is connected to WiFi without config file). This way the ESP will not reconnect to the WiFi AP